api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite==0.19.0
pydantic==2.11.3
google-generativeai==0.3.2
gemini-webapi==1.10.2
browser-cookie3==0.20.1
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0